+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.43.0] - 2022-02-04
+### Added
+- basic Tailwind set-up
+- Tabs and Flickity Vue components
+
 ## [2.42.2] - 2022-01-31
 ### Fixed
 - always fill in purpose in JIRA request in order